close first  REPL after second REPL for same file
added code to close the first instance of a REPL console for a given
file when a second REPL instance launched. Added a bool setting
'close_dup_tab' to turn feature on and off. Also added the source
filename to the REPL tab (view) title.

Also added code to open REPL consoles in the last group, as the last
view. this can be configured with the setting 'open_in_last_group'